requirements.txt
pillow~=12.0.0
PyQt6~=6.10.1
opencv-python~=4.12.0.88
numpy~=2.2.6
setuptools~=80.9.0
pynput~=1.8.1
mss~=10.1.0